# api-interna/openapi/openapi_signature_v1.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Interna de PortaFIB que ofereix serveis de firma web.",
    "description" : "Conjunt de Serveis REST de PortaFIB per atendre peticions de firma a través de web de PortaFIB",
    "contact" : {
      "name" : "Departament de Govern Digital a la Fundació Bit",
      "url" : "http://governdigital.fundaciobit.org",
      "email" : "otae@fundaciobit.org"
    },
    "license" : {
      "name" : "European Union Public Licence (EUPL v1.2)",
      "url" : "https://joinup.ec.europa.eu/sites/default/files/custom-page/attachment/eupl_v1.2_es.pdf"
    },
    "version" : "1.0-SNAPSHOT"
  },
...
  "servers" : [ {
    "url" : "/portafibapi/interna"
  }, {
    "url" : "http://localhost:8080/portafibapi/interna"
  }, {
    "url" : "https://dev.caib.es/portafibapi/interna"
  }, {
    "url" : "https://proves.caib.es/portafibapi/interna"
  }, {
    "url" : "https://se.caib.es/portafibapi/interna"
  }, {
    "url" : "https://www.caib.es/portafibapi/interna"
  } ],
...
  "paths" : {
...
    "/secure/directsignatureonweb/v1/getDocumentaryTypes" : {
      "get" : {
        "tags" : [ "DirectSignatureOnWeb v1" ],
        "summary" : "Retorna una llista dels Tipus Documentals disponibles en el servidor: tipus documentals base, tipus documentals de l'entitat i tipus documentals de l'usuari aplicació",
        "operationId" : "getDocumentaryTypes",
        "parameters" : [ {
          "name" : "language",
          "in" : "query",
          "description" : "Idioma en que s'han de retornar les dades(Només suportat 'ca' o 'es')",
          "schema" : {
            "type" : "string",
            "default" : "ca"
          },
          "examples" : {
            "Català" : {
              "description" : "Català",
              "value" : "ca"
            },
            "Castellano" : {
              "description" : "Castellano",
              "value" : "es"
            }
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Paràmetres incorrectes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "401" : {
            "description" : "No Autenticat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "403" : {
            "description" : "No autoritzat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "500" : {
            "description" : "Error no controlat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "200" : {
            "description" : "Operació realitzada correctament",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentaryTypes"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/directsignatureonweb/v1/getLanguages" : {
      "get" : {
        "tags" : [ "DirectSignatureOnWeb v1" ],
        "summary" : "Retorna els idiomes disponibles.",
        "operationId" : "getLanguages",
        "parameters" : [ {
          "name" : "language",
          "in" : "query",
          "description" : "Idioma en que s'han de retornar les dades(Només suportat 'ca' o 'es')",
          "schema" : {
            "type" : "string",
            "default" : "ca"
          },
          "examples" : {
            "Català" : {
              "description" : "Català",
              "value" : "ca"
            },
            "Castellano" : {
              "description" : "Castellano",
              "value" : "es"
            }
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Paràmetres incorrectes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "401" : {
            "description" : "No Autenticat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "403" : {
            "description" : "No autoritzat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "500" : {
            "description" : "Error no controlat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "200" : {
            "description" : "Operació realitzada correctament",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Languages"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/directsignatureonweb/v1/getProfiles" : {
      "get" : {
        "tags" : [ "DirectSignatureOnWeb v1" ],
        "summary" : "Retorna els perfils de firma.",
        "operationId" : "getProfiles",
        "parameters" : [ {
          "name" : "language",
          "in" : "query",
          "description" : "Idioma en que s'han de retornar les dades(Només suportat 'ca' o 'es')",
          "schema" : {
            "type" : "string",
            "default" : "ca"
          },
          "examples" : {
            "Català" : {
              "description" : "Català",
              "value" : "ca"
            },
            "Castellano" : {
              "description" : "Castellano",
              "value" : "es"
            }
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Paràmetres incorrectes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "401" : {
            "description" : "No Autenticat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "403" : {
            "description" : "No autoritzat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "500" : {
            "description" : "Error no controlat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "200" : {
            "description" : "Operació realitzada correctament",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Profiles"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
...
    "/secure/signatureonserver/v1/getDocumentaryTypes" : {
      "get" : {
        "tags" : [ "SignatureOnServer v1" ],
        "summary" : "Retorna una llista dels Tipus Documentals disponibles en el servidor: tipus documentals base, tipus documentals de l'entitat i tipus documentals de l'usuari aplicació",
        "operationId": "getDocumentaryTypes",
        "parameters" : [ {
          "name" : "language",
          "in" : "query",
          "description" : "Idioma en que s'han de retornar les dades(Només suportat 'ca' o 'es')",
          "schema" : {
            "type" : "string",
            "default" : "ca"
          },
          "examples" : {
            "Català" : {
              "description" : "Català",
              "value" : "ca"
            },
            "Castellano" : {
              "description" : "Castellano",
              "value" : "es"
            }
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Paràmetres incorrectes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "401" : {
            "description" : "No Autenticat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "403" : {
            "description" : "No autoritzat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "500" : {
            "description" : "Error no controlat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "200" : {
            "description" : "Operació realitzada correctament",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentaryTypes"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/signatureonserver/v1/getLanguages" : {
      "get" : {
        "tags" : [ "SignatureOnServer v1" ],
        "summary" : "Retorna els idiomes disponibles.",
        "operationId": "getLanguages",
        "parameters" : [ {
          "name" : "language",
          "in" : "query",
          "description" : "Idioma en que s'han de retornar les dades(Només suportat 'ca' o 'es')",
          "schema" : {
            "type" : "string",
            "default" : "ca"
          },
          "examples" : {
            "Català" : {
              "description" : "Català",
              "value" : "ca"
            },
            "Castellano" : {
              "description" : "Castellano",
              "value" : "es"
            }
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Paràmetres incorrectes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "401" : {
            "description" : "No Autenticat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "403" : {
            "description" : "No autoritzat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "500" : {
            "description" : "Error no controlat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "200" : {
            "description" : "Operació realitzada correctament",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Languages"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/signatureonserver/v1/getProfiles" : {
      "get" : {
        "tags" : [ "SignatureOnServer v1" ],
        "summary" : "Retorna els perfils de firma.",
        "operationId": "getProfiles",
        "parameters" : [ {
          "name" : "language",
          "in" : "query",
          "description" : "Idioma en que s'han de retornar les dades(Només suportat 'ca' o 'es')",
          "schema" : {
            "type" : "string",
            "default" : "ca"
          },
          "examples" : {
            "Català" : {
              "description" : "Català",
              "value" : "ca"
            },
            "Castellano" : {
              "description" : "Castellano",
              "value" : "es"
            }
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Paràmetres incorrectes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "401" : {
            "description" : "No Autenticat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "403" : {
            "description" : "No autoritzat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "500" : {
            "description" : "Error no controlat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestExceptionInfo"
                }
              }
            }
          },
          "200" : {
            "description" : "Operació realitzada correctament",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Profiles"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "RestExceptionInfo" : {
        "required" : [ "errorMessage" ],
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "integer",
            "description" : "Codi intern de l'error. Si l'Aplicació no gestiona codis d'error llavors val null.",
            "format" : "int32"
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Missatge de l'error"
          },
          "stackTrace" : {
            "type" : "string",
            "description" : "Stacktrace de l'excepció"
          },
          "stackTraceCause" : {
            "type" : "string",
            "description" : "Stacktrace de l'excepció causant de l'error si n'hi hagués."
          },
          "field" : {
            "type" : "string",
            "description" : "Indica el camp en que hi ha un error de validació."
          }
        },
        "description" : "Estructura de dades utilitzada per passar informació d'un error"
      },
...
      "KeyValue" : {
        "required" : [ "key", "value" ],
        "type" : "object",
        "properties" : {
          "key" : {
            "type" : "string",
            "description" : "Clau del valor"
          },
          "value" : {
            "type" : "string",
            "description" : "Valor de la clau"
          }
        },
        "description" : "Objecte per afegir informació addicional sobre la firma o el firmant"
      },
      "DocumentaryType" : {
        "required" : [ "documentType", "documentTypeBase", "name" ],
        "type" : "object",
        "properties" : {
          "documentType" : {
            "type" : "integer",
            "description" : "Identificador de Tipus Documental",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "description" : "Nom del Tipus Documental"
          },
          "documentTypeBase" : {
            "type" : "integer",
            "description" : "Identificador de Tipus Documental associat als valors ENI",
            "format" : "int64"
          }
        },
        "description" : "Model de dades d'un tipus documental"
      },
      "DocumentaryTypes" : {
        "type" : "object",
        "properties" : {
          "documentaryTypes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DocumentaryType"
            }
          }
        }
      },
      "Languages" : {
        "type" : "object",
        "properties" : {
          "languages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/KeyValue"
            }
          }
        }
      },
      "Profile" : {
        "required" : [ "code", "description", "name" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Codi del perfil"
          },
          "name" : {
            "type" : "string",
            "description" : "Nom del perfil en l’idioma elegit."
          },
          "description" : {
            "type" : "string",
            "description" : "Descripció del perfil en l’idioma elegit."
          }
        },
        "description" : "Detalls del Perfil d'un usuari Aplicació"
      },
      "Profiles" : {
        "type" : "object",
        "properties" : {
          "profiles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Profile"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}